Rename util folder and file to utils to match naming convention
